requirements.txt
streamlit
scikit-learn
numpy
pillow
streamlit-drawable-canvas
